+++ packages/react/src/blocks/forms/Select.tsx
@@ -9,6 +9,7 @@ export interface FormSelectProps {
   name?: string;
   value?: string;
   defaultValue?: string;
+  required?: boolean;
 }
 
 class FormSelectComponent extends React.Component<FormSelectProps> {
@@ -20,6 +21,7 @@ class FormSelectComponent extends React.Component<FormSelectProps> {
         key={Builder.isEditing && this.props.defaultValue ? this.props.defaultValue : 'default-key'}
         defaultValue={this.props.defaultValue}
         name={this.props.name}
+        required={this.props.required}
         {...this.props.attributes}
       >
         {options &&