+++ packages/sdks/mitosis.config.cjs
@@ -54,6 +54,17 @@ const SRCSET_PLUGIN = () => ({
 });
 
 /**
+ * @type {Plugin}
+ */
+const FETCHPRIORITY_CAMELCASE_PLUGIN = () => ({
+  code: {
+    pre: (code) => {
+      return code.replace(/fetchpriority=/g, 'fetchPriority=');
+    },
+  },
+});
+
+/**
  * Replaces all uses of the native `Text` component with our own `BaseText` component that injects inherited CSS styles
  * to `Text`, mimicking CSS inheritance.
  * @type {Plugin}
@@ -197,20 +208,6 @@ const filterActionAttrBindings = (json, item) => {
     return blocksAttrs || dynamicRendererAttrs;
   });
 };
-
-const ANGULAR_REMOVE_UNUSED_LINK_COMPONENT_PROP_PLUGIN = () => ({
-  code: {
-    post: (code) => {
-      if (code.includes('<enable-editor')) {
-        code = code.replace('[linkComponent]="linkComponent"', '');
-      }
-      if (code.includes('<block-wrapper')) {
-        code = code.replace('[linkComponent]="linkComponent"', '');
-      }
-      return code;
-    },
-  },
-});
 
 // for fixing circular dependencies
 const ANGULAR_FIX_CIRCULAR_DEPENDENCIES_OF_COMPONENTS = () => ({
@@ -595,7 +592,6 @@ module.exports = {
       typescript: true,
       state: 'class-properties',
       plugins: [
-        ANGULAR_REMOVE_UNUSED_LINK_COMPONENT_PROP_PLUGIN,
         ANGULAR_FIX_CIRCULAR_DEPENDENCIES_OF_COMPONENTS,
         ANGULAR_OVERRIDE_COMPONENT_REF_PLUGIN,
         ANGULAR_COMPONENT_NAMES_HAVING_HTML_TAG_NAMES,
@@ -609,6 +605,7 @@ module.exports = {
     },
     solid: {
       typescript: true,
+      stylesType: 'style-tag',
       plugins: [
         INJECT_ENABLE_EDITOR_ON_EVENT_HOOKS_PLUGIN,
         REMOVE_SET_CONTEXT_PLUGIN_FOR_FORM,
@@ -651,6 +648,7 @@ module.exports = {
       typescript: true,
       plugins: [
         SRCSET_PLUGIN,
+        FETCHPRIORITY_CAMELCASE_PLUGIN,
         INJECT_ENABLE_EDITOR_ON_EVENT_HOOKS_PLUGIN,
         REMOVE_SET_CONTEXT_PLUGIN_FOR_FORM,
       ],
@@ -661,6 +659,7 @@ module.exports = {
       typescript: true,
       plugins: [
         SRCSET_PLUGIN,
+        FETCHPRIORITY_CAMELCASE_PLUGIN,
         REMOVE_SET_CONTEXT_PLUGIN_FOR_FORM,
         () => ({
           json: {
@@ -770,6 +769,7 @@ module.exports = {
     qwik: {
       typescript: true,
       plugins: [
+        FETCHPRIORITY_CAMELCASE_PLUGIN,
         /**
          * cleanup `onMount` hooks
          * - rmv unnecessary ones