+++ packages/sdks/src/blocks/image/component-info.ts
@@ -18,7 +18,7 @@ export const componentInfo: ComponentInfo = {
       name: 'image',
       type: 'file',
       bubble: true,
-      allowedFileTypes: ['jpeg', 'jpg', 'png', 'svg'],
+      allowedFileTypes: ['jpeg', 'jpg', 'png', 'svg', 'webp'],
       required: true,
       defaultValue:
         'https://cdn.builder.io/api/v1/image/assets%2FYJIGb4i01jvw0SRdL5Bt%2F72c80f114dc149019051b6852a9e3b7a',
@@ -156,6 +156,13 @@ export const componentInfo: ComponentInfo = {
       helperText:
         "When child blocks are provided, fit to them instead of using the image's aspect ratio",
       defaultValue: true,
+    },
+    {
+      name: 'highPriority',
+      type: 'boolean',
+      advanced: true,
+      helperText:
+        'Mark this image as high priority compared to other images on the page. This prevents lazy loading of the image and tells the browser to load this image before others on the page.',
     },
     {
       name: 'aspectRatio',