+++ packages/utils/src/translation-helpers.ts
@@ -2,6 +2,7 @@ import { BuilderContent } from '@builder.io/sdk';
 import traverse from 'traverse';
 import omit from 'lodash/omit';
 import unescape from 'lodash/unescape';
+import set from 'lodash/set';
 
 export const localizedType = '@builder.io/core:LocalizedValue';
 
@@ -30,6 +31,108 @@ function unescapeStringOrObject(input: string | Record<string, any>) {
 
   // Return input as is if it's neither a string nor an object
   return input;
+}
+
+function recordValue({
+  results,
+  value,
+  path,
+  instructions,
+  sourceLocaleId,
+}: {
+  results: TranslateableFields;
+  value: any;
+  path: string;
+  instructions: string;
+  sourceLocaleId: string;
+}) {
+  if (Array.isArray(value)) {
+    value.forEach((item, index) => {
+      recordValue({
+        results,
+        value: item,
+        path: `${path}[${index}]`,
+        instructions,
+        sourceLocaleId,
+      });
+    });
+  } else if (typeof value === 'object') {
+    if (value['@type'] === localizedType) {
+      results[path] = {
+        value: value?.[sourceLocaleId] || value?.Default,
+        instructions,
+      };
+    } else {
+      Object.entries(value).forEach(([key, v]) => {
+        recordValue({
+          results,
+          value: v,
+          path: `${path}.${key}`,
+          instructions,
+          sourceLocaleId,
+        });
+      });
+    }
+  }
+}
+
+function resolveTranslation({
+  data,
+  basePath,
+  path,
+  value,
+  translation,
+  transformedMeta,
+  locale,
+}: {
+  data: any;
+  basePath: string;
+  path: string;
+  value: any;
+  translation: any;
+  transformedMeta: Record<string, string>;
+  locale: string;
+}) {
+  if (Array.isArray(value)) {
+    value.forEach((item, index) => {
+      resolveTranslation({
+        data,
+        basePath,
+        path: `${path}[${index}]`,
+        value: item,
+        translation,
+        transformedMeta,
+        locale,
+      });
+    });
+  } else if (typeof value === 'object') {
+    if (value['@type'] === localizedType) {
+      const translatedSymbolInput = translation[`${basePath}${path}`];
+
+      if (!translatedSymbolInput?.value) {
+        return;
+      }
+
+      set(data, path, {
+        ...value,
+        [locale]: unescapeStringOrObject(translatedSymbolInput.value),
+      });
+
+      transformedMeta[`transformed.symbol.data.${path}`] = 'localized';
+    } else {
+      Object.entries(value).forEach(([key, v]) => {
+        resolveTranslation({
+          data,
+          basePath,
+          path: `${path}.${key}`,
+          value: v,
+          translation,
+          transformedMeta,
+          locale,
+        });
+      });
+    }
+  }
 }
 
 export function getTranslateableFields(
@@ -90,18 +193,23 @@ export function getTranslateableFields(
       if (el && el.id && el.component?.name === 'Symbol') {
         const symbolInputs = Object.entries(el.component?.options?.symbol?.data) || [];
         if (symbolInputs.length) {
+          const basePath = `blocks.${el.id}.symbolInput`;
           symbolInputs.forEach(
             ([symbolInputName, symbolInputValue]: [
               symbolInputName: string,
               symbolInputValue: any
             ]) => {
-              results[`blocks.${el.id}.symbolInput#${symbolInputName}`] = {
-                value:
-                  typeof symbolInputValue === 'string'
-                    ? symbolInputValue
-                    : symbolInputValue?.[sourceLocaleId] || symbolInputValue?.Default,
+              if (symbolInputName === 'children') {
+                return;
+              }
+
+              recordValue({
+                results,
+                value: symbolInputValue,
+                path: `${basePath}#${symbolInputName}`,
                 instructions: el.meta?.instructions || defaultInstructions,
-              };
+                sourceLocaleId,
+              });
             }
           );
         }
@@ -138,24 +246,22 @@ export function applyTranslation(
       if (el && el.id && el.component?.name === 'Symbol') {
         const symbolInputs = Object.entries(el.component?.options?.symbol?.data) || [];
         if (symbolInputs.length) {
+          const transformedMeta = {};
+
           symbolInputs.forEach(
             ([symbolInputName, symbolInputValue]: [
               symbolInputName: string,
               symbolInputValue: any
             ]) => {
-              const translatedSymbolInput =
-                translation[`blocks.${el.id}.symbolInput#${symbolInputName}`];
-
-              if (!translatedSymbolInput?.value) {
-                return;
-              }
-
-              const localizedValues =
-                typeof symbolInputValue === 'string'
-                  ? {
-                      Default: symbolInputValue,
-                    }
-                  : symbolInputValue;
+              resolveTranslation({
+                data: el.component?.options?.symbol?.data,
+                basePath: `blocks.${el.id}.symbolInput#`,
+                path: symbolInputName,
+                value: symbolInputValue,
+                translation,
+                transformedMeta,
+                locale,
+              });
 
               this.update({
                 ...el,
@@ -163,7 +269,7 @@ export function applyTranslation(
                   ...el.meta,
                   translated: true,
                   // this tells the editor that this is a forced localized input similar to clicking the globe icon
-                  [`transformed.symbol.data.${symbolInputName}`]: 'localized',
+                  ...transformedMeta,
                 },
                 component: {
                   ...el.component,
@@ -173,12 +279,6 @@ export function applyTranslation(
                       ...(el.component.options.symbol || {}),
                       data: {
                         ...(el.component.options.symbol?.data || {}),
-
-                        [symbolInputName]: {
-                          '@type': localizedType,
-                          ...localizedValues,
-                          [locale]: unescapeStringOrObject(translatedSymbolInput.value),
-                        },
                       },
                     },
                   },