+++ plugins/smartling/src/plugin.tsx
@@ -30,6 +30,33 @@ import stringify from 'fast-json-stable-stringify';
 // translation status that indicate the content is being queued for translations
 const enabledTranslationStatuses = ['pending', 'local'];
 
+function updatePublishCTA(content: any, translationModel: any) {
+  let publishButtonText = undefined;
+  let publishedToastMessage = undefined;
+
+  // establish that it's a job's content entry that we are currently in
+  if (content.modelId === translationModel?.id) {
+    const pluginSettings = appState.user.organization?.value?.settings?.plugins?.get(pkg.name);
+    if (!pluginSettings) {
+      return;
+    }
+
+    const enableJobAutoAuthorization = pluginSettings.get('enableJobAutoAuthorization');
+
+    // if 'enableJobAutoAuthorization' is undefined then assume it to be true and proceed likewise
+    if (enableJobAutoAuthorization === undefined || enableJobAutoAuthorization === true) {
+      publishButtonText = 'Authorize';
+      publishedToastMessage = 'Authorized';
+    } else {
+      publishButtonText = 'Send to Smartling';
+      publishedToastMessage = 'Sent to Smartling';
+    }
+  }
+
+  appState.designerState.editorOptions.publishButtonText = publishButtonText;
+  appState.designerState.editorOptions.publishedToastMessage = publishedToastMessage;
+}
+
 registerPlugin(
   {
     name: 'Smartling',
@@ -50,6 +77,14 @@ registerPlugin(
         type: 'string',
         required: true,
       },
+      {
+        name: 'enableJobAutoAuthorization',
+        friendlyName: 'Authorize Smartling Jobs through Builder',
+        type: 'boolean',
+        defaultValue: true,
+        helperText: 'Allows users to authorize Smartling jobs directly from Builder',
+        requiredPermissions: ['admin'],
+      },
     ],
     onSave: async actions => {
       const pluginPrivateKey = await appState.globalState.getPluginPrivateKey(pkg.name);
@@ -69,14 +104,19 @@ registerPlugin(
         () => {
           return String(appState.designerState.editingContentModel?.lastUpdated || '');
         },
-        async shoudlCheck => {
-          if (!shoudlCheck) {
+        async shouldCheck => {
+          if (!shouldCheck) {
             return;
           }
-          const translationStatus =
-            appState.designerState.editingContentModel.meta.get('translationStatus');
-          const translationRequested =
-            appState.designerState.editingContentModel.meta.get('translationRequested');
+
+          updatePublishCTA(appState.designerState.editingContentModel, getTranslationModel());
+
+          const translationStatus = appState.designerState.editingContentModel.meta.get(
+            'translationStatus'
+          );
+          const translationRequested = appState.designerState.editingContentModel.meta.get(
+            'translationRequested'
+          );
 
           // check if there's pending translation
           const isFresh =
@@ -98,7 +138,9 @@ registerPlugin(
               sourceLocale,
               ''
             );
-            const currentRevision = hash(stringify(translatableFields), { encoding: 'base64' });
+            const currentRevision = hash(stringify(translatableFields), {
+              encoding: 'base64',
+            });
             appState.designerState.editingContentModel.meta.set(
               'translationRevisionLatest',
               currentRevision
@@ -108,7 +150,7 @@ registerPlugin(
                 appState.globalState.showGlobalBlockingLoading('Contacting Smartling ....');
                 await api.updateTranslationFile({
                   translationJobId: lastPublishedContent.meta.translationJobId,
-                  translationModel: getTranslationModel().name,
+                  translationModel: translationModelName,
                   contentId: lastPublishedContent.id,
                   contentModel: appState.designerState.editingModel.name,
                   preview: lastPublishedContent.meta.lastPreviewUrl,
@@ -276,10 +318,15 @@ registerPlugin(
             ...fastClone(content.meta),
             translationStatus: 'local',
             translationJobId,
+            translationBy: pkg.name,
           },
         });
         showJobNotification(translationJobId);
       },
+      isDisabled() {
+        return appState.designerState.hasUnsavedChanges();
+      },
+      disabledTooltip: 'Please publish your changes to add to translation job',
     });
     registerContentAction({
       label: 'Request an updated translation',