+++ packages/react/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @builder.io/react
 
+## 3.2.10
+
+### Patch Changes
+
+- bb4a5fd: Feature: add `webp` support for Image block file uploads.
+- 1f62b28: Fix: Remove `iframely` API key from Embed block logic.
+- Updated dependencies [a5b8810]
+  - @builder.io/sdk@2.2.5
+
+## 3.2.9
+
+### Patch Changes
+
+- 945f26e: Adds the `highPriority` option to the Image block component to ensure eager loading.
+
 ## 3.2.8
 
 ### Patch Changes